s1

+++ Service/Controllers/MoneyController.cs
@@ -2,6 +2,8 @@
 using MongoDB.Driver;
 using Service.Models;
 using Prometheus;
+using Microsoft.Extensions.Caching.Distributed;
+using System.Text.Json;
 
 namespace WebAPIApp.Controllers
 {
@@ -10,14 +12,14 @@ namespace WebAPIApp.Controllers
     public class MoneyController : ControllerBase
     {
         private readonly IMongoCollection<Money> _moneyCollection;
-      
+        private readonly IDistributedCache _cache;
         private static readonly Gauge MoneyCount =
     Metrics.CreateGauge(
         "money_collection_size",
         "");
         private static readonly Counter ErrorsTotal =
           Metrics.CreateCounter("api_errors_total", "");
-        
+
         private static readonly Counter RequestsTotal =
   Metrics.CreateCounter("api_requests_total", "");
 
@@ -31,40 +33,53 @@ namespace WebAPIApp.Controllers
                 });
 
 
-        public MoneyController()
+        public MoneyController(IDistributedCache cache)
         {
-         
+            _cache = cache;
+
             var client = new MongoClient("mongodb://localhost:27017");
-            var database = client.GetDatabase("Numismatic_Club");
-            _moneyCollection = database.GetCollection<Money>("Money");
+            var db = client.GetDatabase("Numismatic_Club");
+            _moneyCollection = db.GetCollection<Money>("Money");
+
+            MoneyCount.Set(_moneyCollection.CountDocuments(_ => true));
             InitializeMetric();
         }
         private async void InitializeMetric()
         {
             var count = await _moneyCollection.CountDocumentsAsync(_ => true);
             MoneyCount.Set(count);
-           
+
         }
+
+       
 
         // GET api/money
         [HttpGet]
         public async Task<ActionResult<Money>> Get()
         {
             RequestsTotal.Inc();
+            string cacheKey = "money_all";
+
+            var cached = await _cache.GetStringAsync(cacheKey);
+            if (cached != null)
+            {
+                var data = JsonSerializer.Deserialize<object>(cached);
+                return Ok(data);
+            }
             var totalCount = await _moneyCollection.CountDocumentsAsync(_ => true);
 
             List<Money> entities;
 
             if (totalCount > 1000)
             {
-                
+
                 entities = await _moneyCollection.Aggregate()
                     .Sample(1000)
                     .ToListAsync();
             }
             else
             {
-               
+
                 entities = await _moneyCollection.Find(_ => true).ToListAsync();
             }
 
@@ -75,38 +90,60 @@ namespace WebAPIApp.Controllers
                 m.Year_of_creation,
                 m.Country
             });
+            var options = new DistributedCacheEntryOptions
+            {
+                AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(5)
+            };
+
+
+            await _cache.SetStringAsync(cacheKey, JsonSerializer.Serialize(result), options);
 
             return Ok(result);
         }
 
         // GET api/money/{id}
         [HttpGet("{id}")]
-       
-
-      
         public async Task<ActionResult<Money_>> Get(string id)
         {
             RequestsTotal.Inc();
 
-            using (RequestLatency.NewTimer())
+
             {
+                if (id.Length != 24)
+                {
+                    ErrorsTotal.Inc();
+                    return NotFound("id состоит из 24 символов");
+                }
+                string cacheKey = $"money_{id}";
+
+                var cached = await _cache.GetStringAsync(cacheKey);
+                if (cached != null)
+                {
+                    var data = JsonSerializer.Deserialize<Money_>(cached);
+                    return Ok(data);
+                }
                 var money = await _moneyCollection.Find(x => x.Id == id).FirstOrDefaultAsync();
 
                 if (money == null)
+                {
+                    ErrorsTotal.Inc();
                     return NotFound();
-
-
+                }
                 var result = new Money_
                 {
                     Name = money.Name,
                     Year_of_creation = money.Year_of_creation,
                     Country = money.Country
                 };
-
+                var options = new DistributedCacheEntryOptions
+                {
+                    AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(5)
+                };
+                await _cache.SetStringAsync(cacheKey, JsonSerializer.Serialize(result), options);
                 return Ok(result);
             }
         }
-        
+
 
         // POST api/money
         [HttpPost]
@@ -114,11 +151,11 @@ namespace WebAPIApp.Controllers
         {
             RequestsTotal.Inc();
 
-            using (RequestLatency.NewTimer())
+
             {
                 try
                 {
-                  
+
                     if (moneyWithoutId == null)
                     {
                         ErrorsTotal.Inc();
@@ -152,7 +189,7 @@ namespace WebAPIApp.Controllers
 
                     await _moneyCollection.InsertOneAsync(money);
 
-                  
+
 
                     var result = new Money_
                     {
@@ -160,12 +197,13 @@ namespace WebAPIApp.Controllers
                         Year_of_creation = money.Year_of_creation,
                         Country = money.Country
                     };
+                    await _cache.RemoveAsync("money_all");
                     MoneyCount.Inc();
                     return CreatedAtAction(nameof(Get), new { id = money.Id }, result);
                 }
                 catch
                 {
-                  
+
                     throw;
                 }
             }
@@ -177,20 +215,26 @@ namespace WebAPIApp.Controllers
 
         {
 
+
             RequestsTotal.Inc();
-            using (RequestLatency.NewTimer())
+            if (id.Length != 24)
             {
-                if (moneyWithoutId == null) 
-                {  
-                  
-                    return BadRequest(); 
+                ErrorsTotal.Inc();
+                return NotFound("id состоит из 24 символов");
+            }
+
+            {
+                if (moneyWithoutId == null)
+                {
+
+                    return BadRequest();
                 }
-                 
+
 
                 var existingMoney = await _moneyCollection.Find(x => x.Id == id).FirstOrDefaultAsync();
                 if (existingMoney == null)
                 {
-                    
+
                     return NotFound();
                 }
 
@@ -207,10 +251,12 @@ namespace WebAPIApp.Controllers
                     x => x.Id == id,
                     updatedMoney
                 );
+                await _cache.RemoveAsync("money_all");
+                await _cache.RemoveAsync($"money_{id}");
 
                 if (result.MatchedCount == 0)
                 {
-                   
+
                     return NotFound();
                 }
 
@@ -231,13 +277,25 @@ namespace WebAPIApp.Controllers
         public async Task<ActionResult> Delete(string id)
         {
             RequestsTotal.Inc();
+
+            if (id.Length != 24)
+            {
+                ErrorsTotal.Inc();
+                return NotFound("id состоит из 24 символов");
+            }
+
             var result = await _moneyCollection.DeleteOneAsync(x => x.Id == id);
 
             if (result.DeletedCount == 0)
             {
-               
+                ErrorsTotal.Inc();
                 return NotFound();
             }
+
+
+
+            await _cache.RemoveAsync($"money_{id}");
+            await _cache.RemoveAsync("money_all");
             MoneyCount.Dec();
             return Ok();
         }
@@ -248,23 +306,30 @@ namespace WebAPIApp.Controllers
         {
             var query = _moneyCollection.AsQueryable();
             RequestsTotal.Inc();
+            string cacheKey = $"money_filter_{comparison}_{year}";
 
+            var cached = await _cache.GetStringAsync(cacheKey);
+            if (cached != null)
+            {
+                var data = JsonSerializer.Deserialize<IEnumerable<Money_>>(cached);
+                return Ok(data);
+            }
             switch (comparison.ToLower())
-                {
-                    case "больше":
-                        query = query.Where(x => x.Year_of_creation > year.Value);
-                        break;
+            {
+                case "больше":
+                    query = query.Where(x => x.Year_of_creation > year.Value);
+                    break;
 
-                    case "меньше":
-                        query = query.Where(x => x.Year_of_creation < year.Value);
-                        break;
+                case "меньше":
+                    query = query.Where(x => x.Year_of_creation < year.Value);
+                    break;
                 default:
                     {
                         ErrorsTotal.Inc();
                         return BadRequest("Значение сравнения разрешено использовать только 'больше' или 'меньше'");
                     }
-                }
-            
+            }
+
             var entities = await Task.Run(() => query.ToList());
             var result = entities.Select(m => new Money_
             {
@@ -272,7 +337,11 @@ namespace WebAPIApp.Controllers
                 Year_of_creation = m.Year_of_creation,
                 Country = m.Country
             });
-
+            var options = new DistributedCacheEntryOptions
+            {
+                AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(5)
+            };
+            await _cache.SetStringAsync(cacheKey, JsonSerializer.Serialize(result), options);
             return Ok(result);
         }
     }
